Merge `or-suggestions.hh` into `suggestions.hh`
No real need for keeping a separate header for such a simple class.

This requires changing a bit `OrSuggestions<T>::operator*` to not throw
an `Error` to prevent a cyclic dependency. But since this error is only
thrown on programmer error, we can replace the whole method by a direct
call to `std::get` which will raise its own assertion if needs be.


Former-commit-id: dcf4f77fac8bc214afd4851b367ce386f0a34658