Move templated functions to `sqlite-impl.hh`
This ensures that use-sites properly trigger new monomorphisations on
one hand, and on the other hand keeps the main `sqlite.hh` clean and
interface-only. I think that is good practice in general, but in this
situation in particular we do indeed have `sqlite.hh` users that don't
need the `throw_` function.


Former-commit-id: 05ec0beb40f5e4e162903570b68837b34811a02d