reword notes on copying
clarify that we are copying between different stores. we have not
introduced that notion or why it would be interesting, but for now it
should be fine to keep it in context of the store directory.

we could move that later to a more detailed explanation of different
store types.


Former-commit-id: 631ca18ee62502dd2e51a8f3c9e2ab656b8382ec